添加Actor类作为场景显示物体。 添加项目基类作为新项目的模板。 减小图形类在项目类中的使用。 修改类结构。

+++ MagicCube/Common/Window/WindowBase.h
@@ -17,8 +17,8 @@ typedef function<void(int, int)> InputKeyHandleEvent;
 typedef function<void(unsigned int)> InputCharHandleEvent;
 typedef function<void(int, int)> ResizeHandleEvent;
 typedef function<void(int, int)> ButtonHandleEvent;
-typedef function<void(int, int)> PositionHandleEvent;
-typedef function<void(int)> ScrollHandleEvent;
+typedef function<void(double, double)> PositionHandleEvent;
+typedef function<void(double)> ScrollHandleEvent;
 
 //窗体基类
 class WindowBase
